Fix flaky GOST 34.10 signature test
The problem was a race between the GOST 34.10 signature and verification
tests. Each registered an EC group, but signature registered it with an
OID and verification without. If during threaded test runs, verificatino
ran first, then signature would fail because there was no OID.

Fix the tests by registering using an OID in both cases. Also fix
EC_Group registration so if the same group is registered with and
without an OID, we update the internal state to use the OID.

Fixes GH #2197