In the cli, init the mlock allocator before entering sandbox
Currently OpenBSD port disables pledge because pledge doesn't have
a flag for mlock. By initializing first, we can still use the allocator.
Later (during shutdown) the munlock call will fail but will just
fail with ENOSYS which is ignored, then we munmap the memory to free
it which will unlock as a side-effect.