Documentation and deprecation updates
Deprecate XTEA, GOST cipher, and Tiger.

GOST cipher is (AFAIK) no longer approved in Russia, so no reason
to keep it.

XTEA is obscure. Only reason it was saved in the last round of
deprecations was its existance in Golang's x/crypto - but now x/crypto
has deprecated it, so ...

Found out today that Tiger has a full-round preimage attack.  Not
practical but still... and with it being almost never used or
implemented, more than enough to push it onto the deprecation list.