Use a smaller AES sbox circuit
The depth of a circuit doesn't matter much for bitsliced software as we are
limited by the execution rate of the core. So a smaller circuit tends to be
faster. Take smaller AES sbox circuits from
http://www.cs.yale.edu/homes/peralta/CircuitStuff/CMT.html which have much
higher depth but smaller gate counts.

On Skylake improves bitsliced AES-128 encryption by 9% and decryption by 6%.
Improvement on Cortex-A53 was smaller (1-3%) but consistent.