Modify calendar interface
The fix for #2465 broke Windows since there it is not allowed
to have a DLL interface function defined on a DLL interface class.
(Why, who the f knows).

As a result we are caught between two platform limitations and it's
not actually possible to have a static function on an class that is
DLL exported... (at least not without adding some distinct annotation
to cover this case)