Avoid a long loop in modular square root with certain composite moduli
There are certain composite integers where it is very hard to find a quadratic
non-residue. In that case our loop to find such a non-residue might take a long
time. Bail out after 256 attempts since in principle for a prime modulus we
expect at most 2 tries to find a non-residue.