Do not accept EC signatures when the public point is invalid
Force g*x+p*y to the point at infinity if g or p is not a valid point.

This does not seem to be a security issue since if you can control the
public point you can just use a regular key and sign whatever you like.
Still, it's odd to accept anything here, and easy to prevent it.

Found by CryptoFuzz reported offline by @guidovranken