Incrase the alignment of all lookup tables
We had done this in an ad-hoc way in the past for eg the AES tables
but given how easy C++'s alignas is to use we should apply it
universally.

Use 256 byte alignment for larger tables since that heads of any
possible issues that arise from machines with large cache lines.