Decode ASN.1 TeletexStrings strings correctly
Or at least less incorrectly.

These were asserted to be UTF-8 subsets which is totally wrong. Now we
decode it as a ISO-8859-1 string. This is also wrong, but it seems the
majority of implementations which used TeletexString actually used
Latin1 instead of T.61 encoding.

RFC 3280 says that implemenations should try to accept either
ISO-8859-1 or T.61 in TeletexString. I'm honestly not sure how or if
it is possible to distinguish them reliably.

TeletexString has been "SHOULD NOT" since 2002 so hopefully there are
very few such certificates still valid.