Fix the OCSP online test
The LetsEncrypt cross-signed certificate expired on March 17th
so the OCSP check fails now.

Not possible to use the new LetsEncrypt intermediate certs for this
test because the new intermediates only do CRL-based revocation.

Switch to using a Digicert intermediate which offers OCSP checking.
This should continue to work until Sep 22, 2030