Modify DES to use a variant of the BMI2 approach
This works portably and is even a little faster, since we are basically
trading a multiply and an AND for a pdep.

Not a complete side channel protection but should be side channel
silent except for those rare processors with cache bank conflicts,
or those which short-circuit multiplication, which are beyond hopeless.