Disable ECKCDSA signing with hash truncation
When truncating a hash, ECKCDSA uses the opposite convention of other
signature algorithms. It is relatively difficult to support this
without extensive work, and given that ECKCDSA is quite obscure and
primarily included for BSI compliance reasons, it seems simpler to
just prohibit creating a signature with hash truncation.

We continue to allow verification of signatures with (incorrect) hash
truncation, to allow verifying signatures created by Botan prior to
this point.

See GH #2742