Allow disabling FFI compat features for Cipher_Mode
If `ffi_compat=false` is passed to `::create_or_throw` or `::create`,
then create the cipher mode object with ffi compatability fixes
disabled (if there are any).

For example, in GCM_Mode, the the size of the update buffer is required
to be a multiple of 64 bytes (block size * scaling factor, default 4)
instead of the more usual default of 16 bytes. This can cause issues
when trying to decrypt content that was not encrypted with botan.