Refactor: AEAD_Mode::set_associated_data[_n]() uses std::span
This changes the public interface of set_associated_data_n() to only
take a std::span. The method was introduced just recently, so we
hopefully won't bother too many users with that API change. As a result,
set_associated_data_n() can be used as the virtual entry point for the
AEAD implementations.