Fix bug handling nonce state in EAX mode
When encrypting, m_nonce_mac would not be cleared after decryption
concluded. This would cause EAX to reject future attempts to set a different AD
because it would think it was in the middle of processing a message.

Similarly the state would not be correctly cleared for decryption if
the MAC was invalid.