Remove blinding from Diffie-Hellman
Investigating #2710 showed that the majority of the reason for Diffie-Hellman
being slower than other implementations was the use of blinding. This converts
short-exponent Diffie-Hellman into full-exponent Diffie-Hellman; fully a factor
of 5x to 16x decrease in performance, depending on the group size. This is
not really worthwhile given our modular exponenetiation is anyway constant time.

This does leak the bit length of the Diffie-Hellman secret key, but unlike say
an ECDSA nonce this does not help an attacker.