Initialize stack array b in generate_prime
OSS-Fuzz 62131 reports an msan violation due to `b` not being
completely initalized. The contract for RNG types is that they always
write the requested output, or otherwise throw. So this report is
probably a bug in cryptofuzz, where the fuzzer driver RNG does not
write the requested amount.

Initializing b here prevents use of uninitialized memory in the event
that the RNG does not uphold the contract.