Prepare for Kyber standardization updates
Use _R3 suffix on the KyberMode to allow us to easily support the
final standardized version. This also makes it a bit more clear to
users that they are using the non-final Kyber.

This deprecates Kyber 90s mode since it is not included in the draft
FIPS, seems to be not nearly as widely implemented as the SHA-3
based version, and supporting it complicates the implementation.