introduce a basic module lifecycle
Modules may now be marked as 'Deprecated' or 'Experimental' in
their <module_info> section. When configuring a build with either
of them, a warning will be issued in the output of ./configure.py.

Experimental modules won't be built by default and may be enabled
either explicitly via --enable-modules or globally (all-at-once) with
--enable-experimental-features.

Deprecated modules are built by default but are scheduled for
removal in a future release of the library.