Introduce Public_Key::raw_public_key_bits()
For most algorithms this returns the canonical byte encoding of the
public key's intrinsic object(s). This is particularly useful for the
algorithms of the PQC NIST competition where the public key encoding
is a simple byte-concatenation of the internal objects.

Some algorithm may not have such a canonical byte serialization and may
throw Not_Implemented in this case (e.g. RSA).

For Key Exchange Schemes this is returning the bare 'public_value'.