Fix bugs in scrypt autotune
Fix a bug due to including the extra 512 bytes. When the user provided
max_mem == 0 (implying no limit desired) this would instead be treated
as a very small limit of 512 bytes, leading to using a small N.

When computing the maximum r/N, scrypt_memory_usage was invoked with
p > 1. Since p contributes (very slightly) to the scrypt memory usage,
this lead to typically using an N half of expected.