Refactor hash2curve
Removes ec_h2c since it was just providing a bit of (internal)
abstraction between EC_Group and pcurves.

Since BSI apparently wants to disable h2c for reasons (#2998), and
there is no longer a module to disable, allow disabling support for it
by removing xmd support. Since XMD is only used for hash to curve,
this is otherwise not noticable to users.