Avoid testing multiple bits in CPUID feature checks
For whatever reason both GCC and Clang generate quite terrible code in
this case. Instead avoid the problem by never setting a bit if one of
the dependency bits is missing, for example on a (hypothetical) core
with AES-NI but SSSE3, just avoid setting AES-NI bit in the first place.