Fix artifact of applying clang-format
Reformating with clang-format caused lines like

if(predicate) // this is sometimes true
  foo();

to be changed into

if(predicate) // this is sometimes true
{
   foo();
}

which isn't quite what matching what we'd like. Move the comment
either up above the if or down into the block, so that the braces
match our normal convension.