Introduce PK_Signer/PK_Verifier ::set_associated_data()
This is a new concept introduced by FIPS 204 and 205 (ML-DSA, SLH-DSA)
where applications get the chance to provide some context that is
incorporated into their signatures. It is conceptually similar to the
associated data in an AEAD, therefore it behaves similarly in the
Signer/Verifier.

Note that algorithms that don't support AD, are supposed to always throw
when an application calls set_associated_data() on them. There is also a
predicate function is_valid_associated_data_length() for applications to
generically check for the support of it.

Co-Authored-By: Fabian Albert <fabian.albert@rohde-schwarz.com>