Support asymmetric RSA keys on TPM 2.0
Along with the actual asymmetric algorithm support this adds a number
of required auxiliaries. Most notably, TPM2::HashFunction is an internal
class exposing the TPM's hashing functionality as a Botan::HashFunction.
'Restricted' keys must use this to obtain a token from the TPM proofing
that the data was hashed (and validated) by the TPM.

The RSA adapter is stashed into the 'tpm2_rsa' submodule so that RSA
support can be disabled at compile time. To facilitate the addition of
ECC keys, an abstract TPM2::PrivateKey class is always part of the main
TPM2 module.

Asymmetric keys can be created, loaded, persisted and evicted as needed.

Finally, asymmetric keys (on the TPM) may now be used to establish
authenticated sessions.

Co-Authored-By: René Meusel <rene.meusel@rohde-schwarz.com>