FIX: `ninja tests` fails to build for 'shared'
When building a specific target (e.g. 'tests'), the job dependency pointed
to 'libbotan-3.so.6'. This target will explicitly build just 'libbotan-3.so.6'
but fail to set up the required symlinks 'libbotan-3.so' and '.so.6.6.0'. As
a result, the linker command with `-lbotan-3` will fail to find the shared
object.