Initial refactoring of speed cmdlet
The file speed.cpp is getting quite large and contorted; moving to a
model where the individual performance tests are classes allows
breaking things up into several files in a natural way.

The big remaining areas yet to be addressed are the public key
benchmarks and the generic benchmarks for block ciphers, hashes, etc.
These will be done in followup PRs.