Workaround apparent libstdc++ bug in Ubuntu 24.04
The iterator checks failed when we attempted to use first to extract
ideal_granularity bytes out of dummy_buffer, but on testing it failed
exactly when ideal_granularity was 256 bytes, the same size as
dummy_buffer itself.

Increase the size and add an assert that the buffer is large enough.