Modernize GHASH used in GCM and GMAC
Most notably, this lifts the limitation that GHASH::update() requires
block-aligned data input by transparently caching non-aligned data
using the AlignmentBuffer helper. With that, the GCM mode of operation
can stream-process data that is not block aligned.

The GMAC used to contain an AlignmentBuffer to handle the same limitation
described above. This was removed, as GHASH now handles non-aligned data
transparently.

Finally, this cleans up the state management of GHASH as much as possible
and removes a few public interfaces.