Deprecate a few more obscure and/or obsolete algorithms [ci skip]
MD5 held on for a long time due to needing it for TLS 1.0, but that's
not a factor anymore.

X9.19 MAC was obsolete 20 years ago

The original Keccak proposal is only, as far as I know, used in Ethereum.
We only implement it because the code was added during the SHA-3 competition,
prior to the NIST changes.