Add a generic pcurves instance
Unlike the existing system which instantiates everything at compile time based
on the curve parameters, this allows choosing the parameters at runtime. This
can be used for application specific curves, or when it is desirable to not use
the dedicated modules for binary size reasons.

Comparing the three approaches (pcurves, pcurves_generic, BigInt) for some
common EC operations (keygen, ECDSA sign, ECDSA verify, ECDH agreement).  All
numbers are operations/second for the brainpool256r1 curve.

| impl     | keygen   | sign     | verify   | agree    |
| -------- | -------- | -------- | -------- | -------- |
| pcurves  | 25600    | 17200    | 10800    | 7700     |
| generic  | 12800    | 10200    | 5800     | 4300     |
| BigInt   | 11900    | 8600     | 4000     | 2600     |

The advantage of the curve-specific implementation is clear, but this module has
performance at least as good as the old BigInt-based EC_Point, and with a
significantly improved situation regarding side channels.

This implementation does not support all curves supported by EC_Point; namely
curves with any of the following properties are not supported:

* Curves with fields larger than 521 bits
* Curves with a order of different bitlength than the field
* Curves with primes not congruent to 3 modulo 4
* Curves with a cofactor

In general creating an application specific curve with any such parameters is
deprecated, and will be removed in Botan4. In the meantime, such curves are
handled by the old BigInt elliptic curve implementation.

Notably secp224r1 is not supported by pcurves_generic, as it uses a prime which
is congruent to 1 modulo 4. Support for secp224r1 requires the dedicated
pcurves_secp224r1 module.