In PK_Decryptor::decrypt_or_random return a fake PMS matching expected content bytes
The decrypt_or_random functions are used to hide PKCS1v1.5 oracle attacks. The
variant that checks content bytes of the message returned a random string in the
event of failure, but this random string did not conform to the content byte
requirements; thus if the application redundantly checked the decrypted message
afterwards, it might still create an oracle.

This didn't affect TLS since after decryption it does not examine the premaster
further, it just feeds the premaster into the KDF.