Add PrimeOrderCurve::mul2_setup_g
The common case for a 2-ary multiplication is one of the points is the
group generator. If we know this is the case we can precompute the
relevant multiples of g in advance of knowing the other point. (This
optimizsation is not implemented in this PR, left for future work)