Avoid using power_mod within the library
Due to historical design `power_mod` cannot make any assumptions about its inputs,
leading to worst case performance. In both remaining uses of `power_mod`, the modulus
in question is public so we can can avoid a constant-time setup for the Barrett
reduction. Also in both cases the exponent is public, so we do not need to use
constant time table lookups.