Don't bother with loop blocking for carry chains with constant lengths
As long as the iteration bound is a compile time constant, and not too large [*],
both GCC and Clang will completely unroll a loop calling word_add or word_sub.

Reformulate word_add so that it can use dword if that is helpful. Unfortunately,
GCC generates better code with dword but Clang only recognizes our current
pattern for a carry chain, so we choose at compile time based on `__clang__`.

Only notable regression is a 3% decrease for ECDSA verification with secp256r1
using GCC. I'm not sure of the cause. Everything else (ECDSA, ECDH, RSA, DH) is
either unchanged, or improved. Improvements generally were between 1 and 7%.
Biggest improvements seen with Clang were with ECDH secp256r1 (7%) and ECDSA
verification secp256r1 (9%), and with GCC ECDH secp521r1 (7%) and ECDSA
verification brainpool384r1 (9%)

[*] Measured 17 for GCC and 32 for Clang, but this probably varies based on
a variety of factors.