Workaround odd performance regression with GCC in redc_crandall
The cause of the regression is unclear but removing this asm block reliably
causes a large (up to 20%) performance regression with secp256k1 operations.

Previously this asm block (or something very close to it) was called from
bigint_monty_maybe_sub. Removing the asm block there caused the regression.
Weirdly no other caller of bigint_monty_maybe_sub, even with secp256k1, seemed
harmed by the removal of the asm.