Add named pseudo-constructors for Ed25519_PrivateKey
The behavior of Ed25519_PrivateKey constructor taking a bytestring varies
depending on the length. This behavior is long-standing but was not clearly
documented.

Add new from_seed and from_bytes functions which verify that the input is
of the expected length and then pass it to the pre-existing constructor.

GH #4701