Split up versioning info
Add runtime getters for fields which were previously only available through a header.

Move the VC revision field to a new internal header that is sparingly included. We still
have to keep the existing BOTAN_VERSION_VC_REVISION macro, but now it is *only* set for
releases, not any git workspace build. This means we no longer need the --no-store-vc-rev
hack that is used to avoid compiler cache misses.

GH #3861