Introduce an _oid user-defined literal
In contrast to a simple constructor call to OID("2.5.29.19"), this parses
the OID into a std::array<> at compile-time and just allocates it at runtime.
This saves a lot of tiny allocations for the string operations and plenty of
calls to to_u32bit(). Especially, as some OIDs are parsed repeatedly at run-
time. See extension_from_oid() and Certificate_Extension::static_oid()