In aes_vperm distinguish between shuffle operations
Some shuffles are always executed with indices in range, and others we rely on
the output of the shuffle being zero if the index is out of range.

For x86 and ARM there is no difference, but for VMX and LSX the masked variant
is slower, so only use that when it is actually required.