For SIMD_4x32 require SSSE3 rather than SSE2
For many operations we require or at least could make good use of pshufb and
palignr, which are not available in SSE2. Equivalent instructions are already
available in the other supported SIMD ISAs.

This does prevent using the previously SSE2-only codepaths on CPUs which support
SSE2 but not SSSE3. This includes various processors like the Intel Pentium4 and
AMD K10. Any such processors are at this point at least 15 years old, and likely
in very little use.