Update EC_Group::PEM_encode to handle upcoming removal of explicit curves
A very long time ago, an unfortunate decision was made that this function should
always return the explicit encoding. This directly conflicts with the desire to
remove such encodings entirely in Botan4, if at all practical.

Add an optional argument that allows specifying the encoding. And update the
new-in-3.8 FFI function so that, there, the named curve encoding is always used.