Generate ec_named.cpp and relevant pcurves files from gen_ec_groups.py
This single script will also generate stub/starter pcurves configurations, if
`Impl = pcurve` is set in ec_groups.txt and the source file does not already exist.

This adds Jinja2 as a dependency to gen_ec_groups.py, since it's a lot more
convenient to template with something a bit more featured than what Python
provides out of the box.

Co-Authored-By: René Meusel <rene.meusel@rohde-schwarz.com>