Reduce overhead from X.509 RPKI tests
These tests are verifying extensions in certain situations but first
we verify the signatures which in practice is 90% of the cost of this
test. Given this, where possible use secp192r1 instead of secp256r1;
this reduces the overall runtime of the test by about a factor of 2.

Fix configuration for Ed25519; it needs SHA-512 here.