Support aarch64 in the strubtest GDB extension
On aarch64 the stack handling works somewhat different. First of all, it
grows 'downward' (towards lower addresses). And second, the stack size is
generally not determined at the beginning of the function, because of that.

To obtain a handle to the stackframe of a running function, we need to break
at the function's 'ret' instruction. There, the stack pointer is updated to
however many bytes are being stored on the stack by the running function.