Improve string<->span conversions
This is a somewhat regular need but was not well supported. Add a new
internal header mem_utils.h with some conversion helpers.

The cast_char_ptr_to_uint8 functions in mem_ops.h are no longer used and are now
only declared if BOTAN_IS_BEING_BUILT is not defined.