Run a full clang-tidy scan on every PR
Previously to handle the fact that clang-tidy is very slow, it was only run on
.cpp files which changed in the PR. However this could miss a situation where a
header was changed and that change causes a clang-tidy warning in a source file
which was otherwise unmodified. Exactly this situation occured in #4255

This is not required anymore thanks to the clang-tidy caching added in #4920 so
switch to just always running a full scan and trust that the cache will
(typically) make this operation reasonably fast.