Base58 optimizations
Just applying some tricks already used elsewhere. 6x improvement in encoding speed.

When mapping the integer [0,58) to a character use the offset trick used in base64

Instead of doing a long integer division by 58 over and over, divide by the
largest power of 58 that will fit into a word, then use word-sized divisions
to extract the digits out of that remainder. Similarly when decoding, multiply
by a power of 58 to reduce the number of long integer multiplications.