New implementation of the Jacobi function
The previous implementation is kind of odd and doesn't match any reference I can
find. Rewrite it to follow (more or less) the presentation in the Handbook of
Applied Cryptography, though avoiding the use of recursion.

This implementation seems to be 2 to 5 times faster than the previous algorithm,
depending on parameter sizes.