Add explicit algorithm for variable-time division of 2^k / x
Currently this just uses the standard Knuth long division algorithm, with a bit
of special casing for the numerator. This leads to a slight improvement for
setup of Barrett reduction with public moduli, about 20% for smaller moduli
(under 512 bit) though very small (~3%) for 2048 bits or higher. It's possible
there is a better algorithm for this case.