Refactor: clarify padding specification in Keccak
The NIST FIPS 202 document specifies these paddings in big-endian notation,
within the implementation we need it in little-endian. Also, the paddings are
of variable bit-lengths that are not byte-aligned.

Additionally to the algorithm-specific padding (e.g "01" for SHA-3) the Keccak
permutation itself also adds a variable-length padding (named "pad10*1" in the
NIST document) to align the entire input with the algorithm's byte rate.

This is an attempt to clarify what's going on here and also some preparatory
work to let both the Keccak[c] and Ascon permutation share some code.