Add FFI view functions for mp type
The function botan_mp_to_hex was particularly a problem since the length of the
buffer is not specified by the caller *and* the documented upper bound was
incorrect (GH #5129) making overflows quite a likely issue.

Correct the documented upper bound, and add view functions for hex, decimal, and
binary encoding of botan_mp_t so that in the future callers can avoid the
situation entirely.