Change FFI SRP6 to use fixed length outputs for integer values
The test added in #5135 (quite reasonably) assumed that the SRP6 outputs are
always exactly the same size as the group element, which would occasionally
(~ 1/256 of the time) cause a failure if the integer had a leading zero byte.

Change the FFI layer to zero pad as required so that these short outputs cannot
occur. This is anyway a more consistent and less surprising behavior.