Enable the clang-tidy hicpp-no-assembler and portability-simd-intrinsics warnings
Which is not to suggest that either inline asm or SIMD intrinsics are prohibited
or even discouraged. However being required to explicitly call them out with
comments is helpful for review purposes. Also for SIMD it helps clearly mark
areas in need of improvement; in general the preference is for SIMD operations
to be implemented in the utils/simd headers, or at least invoked via a few
inlined functions when we are calling some algorithm specific intrinsic (for
example in the SHA-NI and SM4-NI code). The direct use of SIMD intrinsics in
large blocks of code, as seen in for example IDEA SSE2 implementation or
the AES-NI key schedules, is definitely a smell.