Use a new strategy for blinding in ECDSA signing
Previously the inversion and the computation of x*r+m were independently
blinded. However this had the disadvantage of requiring knowledge of the inverse
of the second blinding factor. By blinding both computations "together" with the
same blinding factor, several multiplications in the scalar field are avoided.
However more critically, it avoids having to perform an extra inversion during
the setup phase, which matters a lot for one-shot signers since this extra
inversion can easily cost 100s of thousands of cycles (depending on the curve).