Fix: x509 name constraint prefix handling
Fix an inconsistency with OpenSSL and strongSwan
where Botan checked whether the name constraints
are present in the subject DN while the other
implementations matched the constraint as a prefix.

This required moving the x500_name_cmp function to
the internal x509_utils.h header since it also
needs to be used by name_constraint.cpp.