Add a test for encoding of P-521 keys with a leading zero byte.
There was a bug in ECDSA key encoding (GH #4110) which would have been
caught by a stability test, but there already *was* a stability test
(GH #3269). But there was not a test case where the generated key had
a leading zero byte, which would have exposed the bug. Add that test case.